Refactor classes into packages
The project has too many classes that are not in packages.

It will be easier to understand if there were more packages to categorise the classes.

Let's refactor the classes in this structure:
*skeltal:
  *task
    *expense
    *types
*skeltalgui
