+++ src/main/Data/skeltal.txt
@@ -1,1 +1,5 @@
-T | 0 | quick
+T | 1 | akljsjaj
+E | 0 | quick | anytime
+E | 0 | nice one | 12-12-19
+D | 0 | adasd | 12111999
+E | 0 | akjsdhaks | 12-12-1999
