+++ src/main/java/skeltalgui/DialogBox.java
@@ -1,3 +1,5 @@
+package skeltalgui;
+
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
