+++ src/main/java/skeltalgui/Main.java
@@ -1,3 +1,5 @@
+package skeltalgui;
+
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
